# dev_core/api/openapi.yaml
openapi: 3.0.3
info:
  title: InduFrame Control Plane API
  version: 0.1.0
  description: dev_core 面向 dev_ide 的控制面契约。
servers:
  - url: /api/v1
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [user]
      summary: listUsers
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      operationId: createUser
      tags: [user]
      summary: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /users/{userId}:
    put:
      operationId: updateUser
      tags: [user]
      summary: updateUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      operationId: deleteUser
      tags: [user]
      summary: deleteUser
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /users/{userId}/password:
    put:
      operationId: changeUserPassword
      tags: [user]
      summary: changeUserPassword
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        default:
          description: 失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    agentTokenAuth:
      type: http
      scheme: bearer
      bearerFormat: NodeAgentToken
  schemas:
...
    ChangePasswordRequest:
      type: object
      properties:
        oldPassword: { type: string, format: password }
        newPassword: { type: string, format: password, minLength: 8 }
...
    UserCreateRequest:
      type: object
      additionalProperties: false
      description: 仅组织管理员可创建用户。仅用户名、密码、角色必填，其余为选填高级资料。
      required: [username, password, role]
      properties:
        username: { type: string, minLength: 3, maxLength: 50 }
        password: { type: string, format: password, minLength: 1, description: 非空临时密码；首次登录必须改密，新密码至少8个字符且不同于旧密码，不要求字符组合 }
        email: { type: string, maxLength: 254 }
        phone: { type: string, maxLength: 40 }
        fullName: { type: string, maxLength: 100 }
        gender: { type: string, enum: ['', male, female, other, undisclosed] }
        attributes:
          type: object
          description: 自定义文本资料，最多20项；名称不重复且不使用账号或权限字段名，不参与鉴权。
          maxProperties: 20
          additionalProperties: { type: string, maxLength: 500 }
        role: { type: string, enum: [SYSTEM_ADMIN, PROJECT_ADMIN, OPS_ADMIN] }
        status: { type: string, enum: [active, inactive, suspended] }
    UserUpdateRequest:
      type: object
      additionalProperties: false
      description: 用户名不可修改；不传字段保留，空字符串或空属性对象清空。不能修改自己的角色或状态。
      properties:
        email: { type: string, maxLength: 254 }
        phone: { type: string, maxLength: 40 }
        fullName: { type: string, maxLength: 100 }
        gender: { type: string, enum: ['', male, female, other, undisclosed] }
        attributes:
          type: object
          description: 自定义文本资料，最多20项；名称不重复且不使用账号或权限字段名，不参与鉴权。
          maxProperties: 20
          additionalProperties: { type: string, maxLength: 500 }
        role: { type: string, enum: [SYSTEM_ADMIN, PROJECT_ADMIN, OPS_ADMIN] }
        status: { type: string, enum: [active, inactive, suspended] }
...
    ApiResponse:
      type: object
      required: [code, msg, data, reqId]
      properties:
        code:
          type: integer
        msg:
          type: string
        data:
          nullable: true
        reqId:
          type: string
security:
  - bearerAuth: []